Ajout du test sur l'existance du fichier de données à lire
git-svn-id: https://subversion.renater.fr/g2t/G2T-v3@28 d2c9f5c4-e840-43a4-9a69-01eb24ae821f

+++ CRON/import_absence.php
@@ -18,11 +18,17 @@
 
 	// On charge la table des absences HARPEGE avec le fichier
 	$filename = "../INPUT_FILES_V3/har_absence_$date.dat";
-	$load_affect=mysql_query("LOAD DATA LOCAL INFILE '$filename' INTO TABLE HARPABSENCE CHARACTER SET LATIN1 FIELDS TERMINATED BY ';'");
-	$erreur_requete=mysql_error();
-	if ($erreur_requete!="")
-		echo "LOAD HARPABSENCE FROM FILE => $erreur_requete \n";
-
+	if (!file_exists($filename))
+	{
+		echo "Le fichier $filename n'existe pas !!! \n";
+	}
+	else
+	{
+		$load_affect=mysql_query("LOAD DATA LOCAL INFILE '$filename' INTO TABLE HARPABSENCE CHARACTER SET LATIN1 FIELDS TERMINATED BY ';'");
+		$erreur_requete=mysql_error();
+		if ($erreur_requete!="")
+			echo "LOAD HARPABSENCE FROM FILE => $erreur_requete \n";
+	}
 	echo "Fin de l'import des absences HARPEGE " . date("d/m/Y H:i:s") . "\n";
 
 ?>
