+++ CRON/import_structure.php
@@ -16,47 +16,53 @@
 	//			on update les infos
 
 	$filename = "../INPUT_FILES_V3/har_structures_$date.dat";
-	$fp = fopen("$filename","r");
-	while (!feof($fp))
+	if (!file_exists($filename))
 	{
-		$ligne = fgets($fp); // lecture du contenu de la ligne
-		if (trim($ligne)!="")
+		echo "Le fichier $filename n'existe pas !!! \n";
+	}
+	else
+	{
+		$fp = fopen("$filename","r");
+		while (!feof($fp))
 		{
-	//		echo "Ligne = $ligne \n";
-			$ligne_element = explode(";",$ligne);
-			$code_struct = $ligne_element[0];
-			$nom_long_struct = $ligne_element[1];
-			$nom_court_struct = $ligne_element[2];
-			$parent_struct = $ligne_element[3];
-			$resp_struct = $ligne_element[4];
-			echo "code_struct = $code_struct   nom_long_struct=$nom_long_struct   nom_court_struct=$nom_court_struct   parent_struct=$parent_struct   resp_struct=$resp_struct \n";
-
-			$sql = "SELECT * FROM STRUCTURE WHERE STRUCTUREID='" . $code_struct . "'";
-			$query = mysql_query($sql);
-			$erreur_requete=mysql_error();
-			if ($erreur_requete!="")
-				echo "SELECT STRUCTURE => $erreur_requete \n";
-			if (mysql_num_rows($query) == 0) // Structure manquante
+			$ligne = fgets($fp); // lecture du contenu de la ligne
+			if (trim($ligne)!="")
 			{
-				$sql = sprintf("INSERT INTO STRUCTURE(STRUCTUREID,NOMLONG,NOMCOURT,STRUCTUREIDPARENT,RESPONSABLEID) VALUES('%s','%s','%s','%s','%s')",
-						mysql_real_escape_string($code_struct),mysql_real_escape_string($nom_long_struct),mysql_real_escape_string($nom_court_struct),mysql_real_escape_string($parent_struct),mysql_real_escape_string($resp_struct));
-			}
-			else
-			{
-				$sql = sprintf("UPDATE STRUCTURE SET NOMLONG='%s',NOMCOURT='%s',STRUCTUREIDPARENT='%s',RESPONSABLEID='%s' WHERE STRUCTUREID='%s'",
-						mysql_real_escape_string($nom_long_struct),mysql_real_escape_string($nom_court_struct),mysql_real_escape_string($parent_struct),mysql_real_escape_string($resp_struct),mysql_real_escape_string($code_struct));
-			}
-			mysql_query($sql);
-			$erreur_requete=mysql_error();
-			if ($erreur_requete!="")
-			{
-				echo "INSERT/UPDATE STRUCTURE => $erreur_requete \n";
-				echo "sql = $sql \n";
+		//		echo "Ligne = $ligne \n";
+				$ligne_element = explode(";",$ligne);
+				$code_struct = $ligne_element[0];
+				$nom_long_struct = $ligne_element[1];
+				$nom_court_struct = $ligne_element[2];
+				$parent_struct = $ligne_element[3];
+				$resp_struct = $ligne_element[4];
+				echo "code_struct = $code_struct   nom_long_struct=$nom_long_struct   nom_court_struct=$nom_court_struct   parent_struct=$parent_struct   resp_struct=$resp_struct \n";
+	
+				$sql = "SELECT * FROM STRUCTURE WHERE STRUCTUREID='" . $code_struct . "'";
+				$query = mysql_query($sql);
+				$erreur_requete=mysql_error();
+				if ($erreur_requete!="")
+					echo "SELECT STRUCTURE => $erreur_requete \n";
+				if (mysql_num_rows($query) == 0) // Structure manquante
+				{
+					$sql = sprintf("INSERT INTO STRUCTURE(STRUCTUREID,NOMLONG,NOMCOURT,STRUCTUREIDPARENT,RESPONSABLEID) VALUES('%s','%s','%s','%s','%s')",
+							mysql_real_escape_string($code_struct),mysql_real_escape_string($nom_long_struct),mysql_real_escape_string($nom_court_struct),mysql_real_escape_string($parent_struct),mysql_real_escape_string($resp_struct));
+				}
+				else
+				{
+					$sql = sprintf("UPDATE STRUCTURE SET NOMLONG='%s',NOMCOURT='%s',STRUCTUREIDPARENT='%s',RESPONSABLEID='%s' WHERE STRUCTUREID='%s'",
+							mysql_real_escape_string($nom_long_struct),mysql_real_escape_string($nom_court_struct),mysql_real_escape_string($parent_struct),mysql_real_escape_string($resp_struct),mysql_real_escape_string($code_struct));
+				}
+				mysql_query($sql);
+				$erreur_requete=mysql_error();
+				if ($erreur_requete!="")
+				{
+					echo "INSERT/UPDATE STRUCTURE => $erreur_requete \n";
+					echo "sql = $sql \n";
+				}
 			}
 		}
+		fclose($fp);
 	}
-	fclose($fp);
-
 	echo "Fin de l'import des structures " . date("d/m/Y H:i:s") . "\n";
 
 ?>
